% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microgliamorphology_functions.R
\name{metadata_columns}
\alias{metadata_columns}
\title{Would you like to get some metadata?}
\usage{
metadata_columns(x, y)
}
\arguments{
\item{x}{is your unprocessed input .csv file}

\item{y}{is your processed output .csv file containing all your metadata neatly organized into separate columns}
}
\description{
This function allows you to extract metadata columns from the UniqueID column.
}
\keyword{UniqueID}
